move i18n to its own folder
Test component and platform translations
