+++ packages/visual-editor/src/editor/TranslatableStringField.tsx
@@ -1,5 +1,5 @@
 import { TranslatableString } from "../types/types.ts";
-import { MsgString, pt } from "../utils/i18nPlatform.ts";
+import { MsgString, pt } from "../utils/i18n/platform.ts";
 import { CustomField, FieldLabel } from "@measured/puck";
 import { resolveTranslatableString } from "../utils/resolveTranslatableString.tsx";
 import React from "react";
