+++ packages/visual-editor/src/internal/components/InternalLayoutEditor.tsx
@@ -20,7 +20,7 @@ import { DevLogger } from "../../utils/devLogger.ts";
 import { YextEntityFieldSelector } from "../../editor/YextEntityFieldSelector.tsx";
 import { loadMapboxIntoIframe } from "../utils/loadMapboxIntoIframe.tsx";
 import * as lzstring from "lz-string";
-import { msg, pt, usePlatformTranslation } from "../../utils/i18nPlatform.ts";
+import { msg, pt, usePlatformTranslation } from "../../utils/i18n/platform.ts";
 import { ClipboardCopyIcon, ClipboardPasteIcon } from "lucide-react";
 
 const devLogger = new DevLogger();
