+++ packages/visual-editor/src/internal/puck/components/DateTimeSelector.tsx
@@ -1,6 +1,6 @@
 import { CustomField, FieldLabel } from "@measured/puck";
 import { Calendar } from "lucide-react";
-import { pt } from "../../../utils/i18nPlatform.ts";
+import { pt } from "../../../utils/i18n/platform.ts";
 
 export const DateTimeSelector: CustomField<string | undefined> = {
   type: "custom",
