+++ packages/visual-editor/src/internal/puck/constant-value-fields/FAQsSection.tsx
@@ -4,7 +4,7 @@ import {
   msg,
   pt,
   usePlatformTranslation,
-} from "../../../utils/i18nPlatform.ts";
+} from "../../../utils/i18n/platform.ts";
 import { useMemo } from "react";
 import { resolveTranslatableString } from "../../../utils/resolveTranslatableString.tsx";
 import { TranslatableStringField } from "../../../editor/TranslatableStringField.tsx";
