+++ packages/visual-editor/src/internal/puck/constant-value-fields/ImageList.tsx
@@ -1,5 +1,5 @@
 import { ArrayField } from "@measured/puck";
-import { pt } from "../../../utils/i18nPlatform.ts";
+import { pt } from "../../../utils/i18n/platform.ts";
 import { GalleryImageType } from "../../../types/types";
 
 export const IMAGE_LIST_CONSTANT_CONFIG = (): ArrayField<
