+++ packages/visual-editor/src/internal/puck/constant-value-fields/TextList.tsx
@@ -3,7 +3,7 @@ import { Plus as PlusIcon, Trash2 as TrashIcon } from "lucide-react";
 import { useDocument } from "../../../hooks/useDocument.tsx";
 import { RichText, TranslatableRichText } from "../../../types/types.ts";
 import { getDisplayValue } from "../../../utils/resolveTranslatableString.tsx";
-import { usePlatformTranslation } from "../../../utils/i18nPlatform.ts";
+import { usePlatformTranslation } from "../../../utils/i18n/platform.ts";
 import { useState } from "react";
 
 const TEXT_LIST_BUTTON_COLOR: string = "#969696";
