+++ packages/visual-editor/src/utils/VisualEditorProvider.tsx
@@ -6,7 +6,7 @@ import { StreamFields } from "../types/entityFields.ts";
 import { TailwindConfigContext } from "../hooks/useTailwindConfig.tsx";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { I18nextProvider } from "react-i18next";
-import { i18nComponentsInstance } from "./i18nComponents.ts";
+import { i18nComponentsInstance } from "./i18n/components.ts";
 import { normalizeLocalesInObject } from "./normalizeLocale.ts";
 
 type AllOrNothing<T extends Record<string, any>> =
