+++ packages/visual-editor/src/utils/i18n/index.ts
@@ -1,0 +1,7 @@
+export { i18nComponentsInstance } from "./components.ts";
+export {
+  i18nPlatformInstance,
+  usePlatformTranslation,
+  msg,
+  pt,
+} from "./platform.ts";
